+++ test/src/blocs/src/internet_connection_checker_bloc_test.dart
@@ -1,6 +1,6 @@
 // Package imports:
 import 'package:flutter_test/flutter_test.dart';
-import 'package:internet_connection_checker/internet_connection_checker.dart';
+import 'package:internet_connection_checker_plus/internet_connection_checker_plus.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 
@@ -8,15 +8,15 @@ import 'package:mockito/mockito.dart';
 import 'package:pczn_dev_kit/pczn_dev_kit.dart';
 import 'internet_connection_checker_bloc_test.mocks.dart';
 
-@GenerateNiceMocks([MockSpec<InternetConnectionChecker>()])
+@GenerateNiceMocks([MockSpec<InternetConnection>()])
 void main() {
-  late MockInternetConnectionChecker checker;
+  late InternetConnection checker;
   late InternetConnectionCheckerBloc bloc;
 
   setUp(() {
     TestWidgetsFlutterBinding.ensureInitialized();
 
-    checker = MockInternetConnectionChecker();
+    checker = MockInternetConnection();
     bloc = InternetConnectionCheckerBloc(connectionChecker: checker);
   });
 
@@ -31,27 +31,27 @@ void main() {
       });
 
       test('update to true', () {
-        bloc.update(InternetConnectionStatus.connected);
+        bloc.update(InternetStatus.connected);
         expect(bloc.state, isTrue);
       });
 
       test('update to false', () {
-        bloc.update(InternetConnectionStatus.disconnected);
+        bloc.update(InternetStatus.disconnected);
         expect(bloc.state, isFalse);
       });
     });
 
     group('hasConnection', () {
       test('returns true', () async {
-        when(checker.hasConnection).thenAnswer((_) async => true);
-        final res = await bloc.hasConnection();
+        when(checker.hasInternetAccess).thenAnswer((_) async => true);
+        final res = await bloc.hasInternetAccess();
         expect(res, isTrue);
         expect(bloc.state, isTrue);
       });
 
       test('returns false', () async {
-        when(checker.hasConnection).thenAnswer((_) async => false);
-        final res = await bloc.hasConnection();
+        when(checker.hasInternetAccess).thenAnswer((_) async => false);
+        final res = await bloc.hasInternetAccess();
         expect(res, isFalse);
         expect(bloc.state, isFalse);
       });