+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Changed
+- Use different package for internet connection checking
+
 ## 1.2.3+6 - 2024-08-26
 ### Added
 - InternetConnectionBloc